Switch to some other ARC bindings functions
The *arc* module, in some combination of recent versions of SWIG, Python
and ARC, does not appear to have some functions reexported in top-level
that were used for many years. This might be because reexport did not
make any sense or because the reexport was a bug. Instead the proper
functions are used from submodules.