Add cluster URL to logs
aCT loggers are actually LoggerAdapter objects that receive the extra
cluster parameter to be used with the custom logging format that uses
that parameter. Since LoggerAdapter cannot be used for the pyarcrest
logger, the cluster URL has to be added to pyarcrest logs in a different
way.

To achieve that, every pyarcrest logger handler is added a filter that
adds the cluster attribute, which then gets used in the log formatting
that already uses the custom format as the handlers are taken from aCT logger.

In the future, the handlers for aCT logger could be added the same
filter instead of using the LoggerAdapter. That would not require the
filter to be added to handlers for pyarcrest explicitly.