Add note about logging difference
The way to impart contextual information to logs in the latest (3.11)
documentation differs from the documentation for the current lowest
supported version (3.6). The difference is that 3.11 creates a log
record copy, modifies the copy and returns it, while the previous
versions (not knowing which is the last version working like that)
modify the record and return boolean on whether the record should be
filtered out or not. This is why there were issues implementing this in
the first place.

The note should be used as a reminder that this might have to be
changed.